BrainEyeProject/requirements.txt
Appium-Python-Client==3.8.0
pytest==8.2.0
pytest-html==3.1.1
selenium==4.9.0